+++ src/helix/endpoints/moderation/mod.rs
@@ -10,6 +10,7 @@ use serde::{Deserialize, Serialize};
 pub mod add_blocked_term;
 pub mod ban_user;
 pub mod check_automod_status;
+pub mod delete_chat_messages;
 pub mod get_banned_users;
 pub mod get_blocked_terms;
 pub mod get_moderators;
@@ -25,6 +26,8 @@ pub use ban_user::{BanUser, BanUserBody, BanUserRequest};
 pub use check_automod_status::{
     CheckAutoModStatus, CheckAutoModStatusBody, CheckAutoModStatusRequest,
 };
+#[doc(inline)]
+pub use delete_chat_messages::{DeleteChatMessagesRequest, DeleteChatMessagesResponse};
 #[doc(inline)]
 pub use get_banned_users::{BannedUser, GetBannedUsersRequest};
 #[doc(inline)]