+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Deprecate `bits_voting_enabled`, `bits_per_vote` and `bits_votes` on `Create Poll` and `Poll`
 * Removed deprecated `limit` on `CreateEventSubSubscription`
 * Removed deprecated `AddBlockedTerm`, `DeleteBlockedTerm`, `DeletePermittedTerm` and `AddPermittedTerm` on `ModerationActionCommand`
+
 ### Added
 
 * Added `Get Chatters` endpoint
@@ -30,6 +31,7 @@
 * Added `vod_offset` to `Clip`s
 * Added `Start A Raid` and `Cancel A Raid` endpoints
 * Added `Send Chat Announcement` endpoint
+* Added `Delete Chat Messages` endpoint
 
 ## [v0.6.1] - 2022-04-29